Fixed some conversion warnings caused by intnum commit.  Missed these due to use of configure.ac's --enable-warnerror, which was set up to disable conversion errors due to flex's warning-prone generated code.  As we no longer use flex, fix configure.ac to not disable conversion errors.
svn path=/trunk/yasm/; revision=1018

+++ libyasm/bytecode.c
@@ -691,7 +691,7 @@ bc_tobytes_data(bytecode_data *bc_data, unsigned char **bufp,
 		break;
 	    case DV_EXPR:
 		if (output_expr(&dv->data.expn, *bufp, bc_data->size,
-				bc_data->size*8, 0,
+				(size_t)(bc_data->size*8), 0,
 				(unsigned long)(*bufp-bufp_orig), sect, bc, 0,
 				1, d))
 		    return 1;
